added existance check to remove

+++ pytest-coverage/run_coverage.py
@@ -15,6 +15,8 @@ save_path = "pytest-coverage"
 Path(save_path).mkdir(parents=True, exist_ok=True)
 
 save_path = os.path.join(save_path, "pytest-coverage.txt")
-os.remove(save_path)
+
+if os.path.exists(save_path):
+    os.remove(save_path)
 with open(save_path, "w") as f:
     f.write(output.decode("utf-8"))
